Description of the Homework
You will take three inputs:  •A File name•  A File layout description  •Fields that we want to be printed. The idea is to navigate on the file, go to the exact position where the data we are interested in is located at, fetch the data, display it and then move to the next data field we are interested in.To get it straight from the start, you are not expected nor allowed to read any array data into the memory. 
〈layout description〉::=〈data description〉|〈data description〉,〈layout description〉〈data description〉::=〈integral data description〉|〈floating data description〉〈integral data description〉::=〈integral type〉〈integral data〉〈floating data description〉::=〈floating type〉〈floating data〉〈integral data〉::=〈integral identifier〉|
〈integral identifier〉[〈atomic〉]|〈integral identifier〉[〈atomic〉][〈atomic〉]〈floating data〉::=〈floating identifier〉|〈floating identifier〉[〈atomic〉]|〈floating identifier〉[〈atomic〉][〈atomic〉]〈integral type〉::=char|uchar|int|uint|long〈floating type〉::=float|double〈integral identifier〉::=〈identifier〉〈floating identifier〉::=〈identifier〉〈query〉::=〈integral data〉|〈floating data〉〈query list〉::=〈query〉|〈query〉,〈query list〉〈atomic〉::=〈natural number〉|〈integral identifier〉〈natural number〉::=0|1|. . .|100000000〈identifier〉::=〈letter〉|〈letter〉〈letternums〉〈letternums〉::=〈letter〉|〈digit〉|〈letter〉〈letternums〉|〈digit〉〈letternums〉〈letter〉::=a|b|. . .|z〈digit〉::=0|1|. . .|9


In the〈layout description〉any〈identif ier〉used to define the size of an array (i.e. appearin a pair of square brackets) will have been defined in a prior data description (prior meansa data description to the left of the usage)
•Any〈atomic〉in a query that is used in an array index has a valid value as far as the indexrange is concerned.

SPECIFICATIONS:
•While your program is run for grading your program and heap space will be limited to anamount that will disable you from reading in all the data and then process it.  Do not eventhink of reading in any array.
•The use of period operators (to access subfields of structures, if you define any) is forbiddenexcept in define macros.

Now assume the input lines to thethe3.cprogram are as follows:
sample.dat
int n, float fiko, int kek, float ali[10], double s[n],long w,uint ka3x1,int p12[n][fiko], double z
fiko,n,   p12[1][n],ali[1],ali[kek],ka3x1,    kek,  w,s[2]
The expected output of the program is:
3.1415933
-104
2.250000
25.628906
882334423
7
1234567901011
13.285714